requirements.txt
requests==2.31.0
imaplib2==3.6
python-dotenv